S12P21A203-44 refactor: 코드 리뷰 반영

+++ backend/src/main/java/com/example/lettering/domain/keyring/service/KeyringServiceImpl.java
@@ -82,7 +82,7 @@ public class KeyringServiceImpl implements KeyringService{
         }
         keyringRepository.saveAll(keyrings);
 
-        int totalPrice = (int) (selectedDesign.getPrice() * request.getQuantity());
+        Long totalPrice =  selectedDesign.getPrice() * request.getQuantity();
 
         Order order = Order.create(
                 user, null, // 🚨 orderNumber는 approval에서 세팅
